[CP-2362] Refactor: Styled elements moved to separate file
