+++ libs/core/messages/components/messages-search-results/messages-search-results.component.tsx
@@ -5,30 +5,24 @@
 
 import React from "react"
 import { FunctionComponent } from "Core/core/types/function-component.interface"
-import styled, { css } from "styled-components"
 import {
+  Actions,
   Col,
   EmptyState,
   LoadingState,
-  Actions,
 } from "Core/__deprecated__/renderer/components/core/table/table.component"
-import { basicAvatarStyles } from "Core/__deprecated__/renderer/components/core/avatar/avatar.component"
-import {
-  backgroundColor,
-  borderColor,
-} from "Core/core/styles/theming/theme-getters"
+import { AvatarSize } from "Core/__deprecated__/renderer/components/core/avatar/avatar.component"
 import useTableScrolling from "Core/__deprecated__/renderer/utils/hooks/use-table-scrolling"
 import { MessagesSearchResultsTestIdsEnum } from "Core/messages/components/messages-search-results/messages-search-results-test-ids.enum"
-import { ResultState } from "Core/contacts/reducers/contacts.interface"
+import { Contact, ResultState } from "Core/contacts/reducers/contacts.interface"
 import { Message } from "Core/messages/dto"
-import Text from "Core/__deprecated__/renderer/components/core/text/text.component"
 import { TextDisplayStyle } from "Core/__deprecated__/renderer/components/core/text/text.component"
 import { defineMessages } from "react-intl"
 import {
   InitialsAvatar,
+  LastMessageText,
   ThreadCol,
   ThreadDataWrapper,
-  LastMessageText,
   WarningIconWrapper,
 } from "Core/messages/components/thread-row.styled"
 import { Settings } from "Core/settings/dto"
@@ -40,44 +34,20 @@ import ThreadRowName from "Core/messages/components/thread-row-name"
 import Dropdown from "Core/__deprecated__/renderer/components/core/dropdown/dropdown.component"
 import { IconButtonWithSecondaryTooltip } from "Core/__deprecated__/renderer/components/core/icon-button-with-tooltip/icon-button-with-secondary-tooltip.component"
 import { IconType } from "Core/__deprecated__/renderer/components/core/icon/icon-type"
-import { AvatarSize } from "Core/__deprecated__/renderer/components/core/avatar/avatar.component"
 import Icon from "Core/__deprecated__/renderer/components/core/icon/icon.component"
 import ButtonComponent from "Core/__deprecated__/renderer/components/core/button/button.component"
 import { DisplayStyle } from "Core/__deprecated__/renderer/components/core/button/button.config"
 import { ElementWithTooltipPlace } from "Core/__deprecated__/renderer/components/core/tooltip/element-with-tooltip.component"
-import ScrollAnchorContainer from "Core/__deprecated__/renderer/components/rest/scroll-anchor-container/scroll-anchor-container.component"
-import { Contact } from "Core/contacts/reducers/contacts.interface"
+import ScrollAnchorContainer
+  from "Core/__deprecated__/renderer/components/rest/scroll-anchor-container/scroll-anchor-container.component"
 import { MessageType } from "Core/messages/constants"
 import { intl } from "Core/__deprecated__/renderer/utils/intl"
 import { Threads } from "Core/messages/components/thread-list.component"
-import ThreadBaseRow from "Core/messages/components/thread-base-row.component"
-
-export const lightAvatarStyles = css`
-  background-color: ${backgroundColor("row")};
-`
-
-export const AvatarPlaceholder = styled.div`
-  ${basicAvatarStyles};
-  margin-right: 1.2rem;
-`
-
-export const SearchResultQueryWrapper = styled.div`
-  padding: 0 3.2rem 1.7rem;
-  border-bottom: solid 0.1rem ${borderColor("list")};
-`
-
-export const SearchTitle = styled(Text)`
-  text-overflow: ellipsis;
-  white-space: nowrap;
-  overflow: hidden;
-`
-export const SearchResultContainer = styled(ThreadBaseRow)`
-  :hover {
-    ${InitialsAvatar} {
-      background-color: ${backgroundColor("row")};
-    }
-  }
-`
+import {
+  SearchResultContainer,
+  SearchResultQueryWrapper,
+  SearchTitle,
+} from "Core/messages/components/messages-search-results/messages-search-results.styled"
 
 const messages = defineMessages({
   searchResultsTitle: {
@@ -86,7 +56,23 @@ const messages = defineMessages({
   dropdownTogglerTooltipDescription: {
     id: "component.dropdownTogglerTooltipDescription",
   },
+  emptyResultsListTitle: {
+    id: "module.messages.emptyResultsListTitle",
+  },
+  emptyResultsListDescription: {
+    id: "module.messages.emptyResultsListDescription",
+  },
 })
+
+const MessagesSearchResultsEmptyState = () => {
+  return (
+    <EmptyState
+      title={messages.emptyResultsListTitle}
+      description={messages.emptyResultsListDescription}
+      data-testid={MessagesSearchResultsTestIdsEnum.Empty}
+    />
+  )
+}
 
 interface MessagesSearchResultProps extends Pick<Settings, "language"> {
   resultsState: ResultState
@@ -111,16 +97,6 @@ const MessagesSearchResults: FunctionComponent<MessagesSearchResultProps> = ({
 }) => {
   const { enableScroll, disableScroll, scrollable } = useTableScrolling()
 
-  const emptyList = () => (
-    <EmptyState
-      title={{ id: "module.messages.emptyResultsListTitle" }}
-      description={{
-        id: "module.messages.emptyResultsListDescription",
-      }}
-      data-testid={MessagesSearchResultsTestIdsEnum.Empty}
-    />
-  )
-
   return (
     <div {...restProps}>
       <SearchResultQueryWrapper>
@@ -136,113 +112,112 @@ const MessagesSearchResults: FunctionComponent<MessagesSearchResultProps> = ({
         data-testid={MessagesSearchResultsTestIdsEnum.Table}
       >
         {resultsState === ResultState.Loaded &&
-          (results.length
-            ? results.map((message) => {
-                const phoneNumber = message.phoneNumber
-                const { id, date, content } = message
-                const contact = getContactByPhoneNumber(message.phoneNumber)
-                const isMessageFailed =
-                  message.messageType === MessageType.FAILED
-                const handleRowClick = () => onRowClick(message)
-                // AUTO DISABLED - fix me if you like :)
-                // eslint-disable-next-line @typescript-eslint/no-unsafe-return
-                const remove = () => removeMessage(id)
-                // AUTO DISABLED - fix me if you like :)
-                // eslint-disable-next-line @typescript-eslint/no-unsafe-return
-                const resend = () => resendMessage(id)
+          (results.length ? (
+            results.map((message) => {
+              const phoneNumber = message.phoneNumber
+              const { id, date, content } = message
+              const contact = getContactByPhoneNumber(message.phoneNumber)
+              const isMessageFailed = message.messageType === MessageType.FAILED
+              const handleRowClick = () => onRowClick(message)
+              // AUTO DISABLED - fix me if you like :)
+              // eslint-disable-next-line @typescript-eslint/no-unsafe-return
+              const remove = () => removeMessage(id)
+              // AUTO DISABLED - fix me if you like :)
+              // eslint-disable-next-line @typescript-eslint/no-unsafe-return
+              const resend = () => resendMessage(id)
 
-                return (
-                  <SearchResultContainer
-                    key={id}
-                    selected={false}
-                    active={false}
-                  >
-                    <Col>
-                      <InitialsAvatar
-                        user={contact}
-                        light={false}
-                        size={AvatarSize.Big}
+              return (
+                <SearchResultContainer key={id} selected={false} active={false}>
+                  <Col>
+                    <InitialsAvatar
+                      user={contact}
+                      light={false}
+                      size={AvatarSize.Big}
+                    />
+                  </Col>
+                  <ThreadCol onClick={handleRowClick}>
+                    <ThreadDataWrapper
+                      sidebarOpened={false}
+                      isMessageFailed={isMessageFailed}
+                    >
+                      <ThreadRowName
+                        contact={contact}
+                        phoneNumber={phoneNumber}
                       />
-                    </Col>
-                    <ThreadCol onClick={handleRowClick}>
-                      <ThreadDataWrapper
-                        sidebarOpened={false}
-                        isMessageFailed={isMessageFailed}
+                      <Time
+                        displayStyle={TextDisplayStyle.Label}
+                        color="secondary"
                       >
-                        <ThreadRowName
-                          contact={contact}
-                          phoneNumber={phoneNumber}
-                        />
-                        <Time
-                          displayStyle={TextDisplayStyle.Label}
-                          color="secondary"
-                        >
-                          {isToday(date)
-                            ? moment(date).format("h:mm A")
-                            : moment(date)
-                                .locale(language ?? "en")
-                                .format("ll")}
-                        </Time>
+                        {isToday(date)
+                          ? moment(date).format("h:mm A")
+                          : moment(date)
+                              .locale(language ?? "en")
+                              .format("ll")}
+                      </Time>
 
-                        <LastMessageText
-                          unread={false}
-                          color="secondary"
-                          displayStyle={TextDisplayStyle.Paragraph4}
-                        >
-                          {content}
-                        </LastMessageText>
-                      </ThreadDataWrapper>
-                      {isMessageFailed && (
-                        <WarningIconWrapper>
-                          <Icon type={IconType.Warning} width={1.6} />
-                        </WarningIconWrapper>
-                      )}
-                    </ThreadCol>
-                    <Col>
-                      <Actions>
-                        <Dropdown
-                          onOpen={disableScroll}
-                          onClose={enableScroll}
-                          toggler={
-                            <IconButtonWithSecondaryTooltip
-                              iconType={IconType.More}
-                              description={
-                                messages.dropdownTogglerTooltipDescription
-                              }
-                              // FIXME: The position based on offset is a sticky. However, this is a quick workaround
-                              //  for buggy overridePosition lib feature
-                              place={ElementWithTooltipPlace.Bottom}
-                              offset={{ left: 15, bottom: 5 }}
-                            />
-                          }
-                        >
-                          {isMessageFailed && (
-                            <ButtonComponent
-                              labelMessage={{
-                                id: "module.messages.messageDropdownResend",
-                              }}
-                              Icon={IconType.Send}
-                              onClick={resend}
-                              displayStyle={DisplayStyle.Dropdown}
-                            />
-                          )}
+                      <LastMessageText
+                        unread={false}
+                        color="secondary"
+                        displayStyle={TextDisplayStyle.Paragraph4}
+                      >
+                        {content}
+                      </LastMessageText>
+                    </ThreadDataWrapper>
+                    {isMessageFailed && (
+                      <WarningIconWrapper>
+                        <Icon type={IconType.Warning} width={1.6} />
+                      </WarningIconWrapper>
+                    )}
+                  </ThreadCol>
+                  <Col>
+                    <Actions>
+                      <Dropdown
+                        onOpen={disableScroll}
+                        onClose={enableScroll}
+                        toggler={
+                          <IconButtonWithSecondaryTooltip
+                            iconType={IconType.More}
+                            description={
+                              messages.dropdownTogglerTooltipDescription
+                            }
+                            // FIXME: The position based on offset is a sticky. However, this is a quick workaround
+                            //  for buggy overridePosition lib feature
+                            place={ElementWithTooltipPlace.Bottom}
+                            offset={{ left: 15, bottom: 5 }}
+                          />
+                        }
+                      >
+                        {isMessageFailed && (
                           <ButtonComponent
                             labelMessage={{
-                              id: "module.messages.messageDropdownDelete",
+                              id: "module.messages.messageDropdownResend",
                             }}
-                            Icon={IconType.Delete}
-                            onClick={remove}
+                            Icon={IconType.Send}
+                            onClick={resend}
                             displayStyle={DisplayStyle.Dropdown}
                           />
-                        </Dropdown>
-                      </Actions>
-                    </Col>
-                    <ScrollAnchorContainer key={id} active={false} />
-                  </SearchResultContainer>
-                )
-              })
-            : emptyList())}
-        {resultsState === ResultState.Empty && emptyList()}
+                        )}
+                        <ButtonComponent
+                          labelMessage={{
+                            id: "module.messages.messageDropdownDelete",
+                          }}
+                          Icon={IconType.Delete}
+                          onClick={remove}
+                          displayStyle={DisplayStyle.Dropdown}
+                        />
+                      </Dropdown>
+                    </Actions>
+                  </Col>
+                  <ScrollAnchorContainer key={id} active={false} />
+                </SearchResultContainer>
+              )
+            })
+          ) : (
+            <MessagesSearchResultsEmptyState />
+          ))}
+        {resultsState === ResultState.Empty && (
+          <MessagesSearchResultsEmptyState />
+        )}
         {resultsState === ResultState.Loading && (
           <LoadingState
             data-testid={MessagesSearchResultsTestIdsEnum.Loading}
